Created __structure.py for class Structure
As with class Token, class Structure has been moved into it's own file
for readability. Right now Structure is quite small but it will get
bigger as it will be what the Parser produces. Each Structure instance
will be a section with section keys. Each structure will eventually be
transformed into a dictionary and combined into the user dictionary
returned to the user.
However, in this commit I've simply moved the Structure class out of
__parser.py