Began refactoring: lexer.py
So the current codebase is kind of a mess, I used a lot of complicated coding practices that made the code look really bad and hard to understand.
I'm changing it up, first things first is getting up the lexer and parser.
Shouldn't take long, it's just organising it.

changelog:
- __init__.py removed library abstractions: going for a load/generate style instead.
- main.py updated some testing for the library
- renamed __token.py to token.py
- Added lexer.py