Added parser and parser testing
The parser can be found parser.py and uses the paring nodes defined in the previous commit.
Current the parser does not make checks to ensure a token is the correct token, it just parses the correct pattern.
This means the parser could parse a KeyNode as:
IDENTIFIER DELIMITER IDENTIFIER
because it does not make specific checks to ensure the token after the DELIMITER is an Expression value.
This will need to be fixed later.

changelog:
- Added parser.py
- Updated main.py to include testing
- Updated __init__.py for testing purposes
- inimod can now parse basic ini config files