Fixed lexer false space identifier bug
When an identifier is used (such as when creating a key) a space is used to separate it and the elements around it.
However, since identifiers can contain spaces, all identifiers would include a space at the end.
In this commit, a patch has been made to ensure identifiers do not have a space on the end.

changelog:
- Fixed bug in lexer to do with identifiers ending with a space.