Added user abstracted functions and comments
User abstracted functions are functions supplied to the user from __init__.py which allows the user to use the library without having to message around using functions from the lexer, parser, binder, tokens, etc...

Instead, the user is given the following functions:
- load(str)->dict : Takes .ini config contents and returns a dictionary
- lex(str)->list : Takes .ini config contents and returns a token list
- parse(list)->list : Takes a token list and returns a node list
- bind(list)->dict : Takes a node list and returns a dictionary
- test(str)->None : Takes .ini contents and return console debugging

I believe these functions will be all that's necessary to use the library.
The functions: lex, parse, and bind, allow the user to bypass the abstraction somewhat without being able to mess with class property causing silly errors.

A lot of commenting has been put on these functions in the hopes users using IDEs will be able to read these comments and understand what the function does, and what it abstracts.